backend/requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Database
mysql-connector-python==8.2.0

# Authentication
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
Flask-Bcrypt==1.0.1

# CORS Support
Flask-CORS==4.0.0

# Password Hashing
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Date/Time Utilities
python-dateutil==2.8.2

# Testing (Optional)
pytest==7.4.3
pytest-flask==1.3.0

# Development Tools (Optional)
flask-migrate==4.0.5
